+++ packages/forms/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @under-control/forms
 
+## 1.5.2
+
+### Patch Changes
+
+- submit fixes
+- Updated dependencies
+  - @under-control/validate@1.5.2
+  - @under-control/inputs@1.5.2
+  - @under-control/core@1.5.2
+
+## 1.5.1
+
+### Patch Changes
+
+- add missing exports
+- Updated dependencies
+  - @under-control/core@1.5.1
+  - @under-control/inputs@1.5.1
+  - @under-control/validate@1.5.1
+
 ## 1.5.0
 
 ### Minor Changes